Edit .gitignore for better optimization
When I created this repository during the Hackathon, I quickly created
the .gitignore file to avoid adding unnecessary files to the git
repository. However, I decided to add everything related to Java
(including various unrelated IDEs). This made it so that the file was
filled with excess files to avoid.

I decided to edit the .gitignore to remove all of the excess files and
re-edit it to exclude only the exe files.