Format all code of the project
This project was originally created as a submission for the Fall
TeenHacks LI 24 Hour Hackathon. Due to this, most of the code was
sloppy written in order to meet the deadline- making it hard to
read.

In order to fix this problem, I formatted every single file java
file (includes every era of the quiz, the main file, the file that
plays music, the question bank, and Ryan's minigame.