Fix drawing bugs for Era 2.
Near the end of competition, the code for the second era of the quiz
got accidently deleted. This meant that Talha had to rapidly redo the
entire code, leading to much more sloppy code.

I removed the multiple for loops and fixed the drawing method of that
era (so that it would actually be called).