Stop adding to correctAnswer in QuizEra70s class
When we initially programmed this class, we created an arrayList to
add all of the correct answers to each of the questions in the era.
However, as development continued we decided to ignore the arrayList
and instead hardcode comparing the Strings for the answers instead.
However, in doing so we made it so that the answers would be constantly
added to the arrayList, making it grow very quickly.

In order to fix this, I moved the correctAnswers method.