Add proper constructor to TeenHack_Era2 class
Looking back at the code, I noticed how much of a mess the
creation of the class was (all of the variables were initialized at
the start of the class instead of being declared and assigned
separately). This is similar to what happened with the QuizEra70s
class.

In order to fix this issue, I created a proper constructor.