Fix overflow bug from creating Minigame class
In fixing up the HackathonQuestionsRyan class, I had changed how the
Minigame class was created. Since the Minigame class extended the
HackathonQuestionsRyan class (needed it for ImageObserver), it would
cause a bug where the program would infinitely swap between the
HackathonQuestionsRyan class and the Minigame class- breaking the
entire program.

In order to fix this, I removed ImageObserver and previous
inheritance.