Improve object creation in Alpha class
The Alpha class was designed as being the main class of the project.
This meant that it was the place were the objects of the other classes
were created. However, due to our lackluster knowledge of instantiating
objects we created objects much more than we needed to (for example,
creating an object under every input method in order to send that input
to that respective class). This caused to program to use a lot more
power than necessary to run.

In order to fix this, I removed all of the repeated instantiating of
the classes and created them once in the main game loop.