Add proper README to project.
Recently, I have been going through my older projects in order to make
them more presentable for colleges (as they might look at my profile).
In order to improve their (and everyone else's)
understanding of this project, I added a proper README to it, listing
out all of the information one would need to know.

It includes a Table of Contents, Context information surrounding the
project and information about it as well.