Add videos to README.md
After Github finally added the ability to add videos to Markdown, I decided to go back and replace some of the bigger gifs on this markdown with much smaller videos. I tried to convert the Custom Text gif to a smaller one, but the image noise was too distracting, so I had to leave it as is. Given that these videos are uploaded now, I can delete the bigger gifs they had replaced and in turn make this repository smaller (even if only slightly).